+++ app/client/src/pages/Editor/WidgetsEditor/MainContainerWrapper.tsx
@@ -3,7 +3,6 @@ import React, { useEffect } from "react";
 import { useSelector } from "react-redux";
 
 import {
-  getCanvasWidth,
   getIsFetchingPage,
   getViewModePageList,
   showCanvasTopSectionSelector,
@@ -39,7 +38,6 @@ import {
 } from "../../../layoutSystems/common/useLayoutSystemFeatures";
 import { CANVAS_VIEWPORT } from "constants/componentClassNameConstants";
 import { MainContainerResizer } from "layoutSystems/common/mainContainerResizer/MainContainerResizer";
-import OverlayCanvasContainer from "layoutSystems/common/WidgetNamesCanvas";
 
 interface MainCanvasWrapperProps {
   isPreviewMode: boolean;
@@ -47,7 +45,7 @@ interface MainCanvasWrapperProps {
   navigationHeight?: number;
   isAppSettingsPaneWithNavigationTabOpen?: boolean;
   currentPageId: string;
-  parentRef: React.RefObject<HTMLDivElement | null>;
+  canvasWidth: number;
 }
 
 const Wrapper = styled.section<{
@@ -125,7 +123,6 @@ function MainContainerWrapper(props: MainCanvasWrapperProps) {
   const { currentPageId, isPreviewMode, shouldShowSnapShotBanner } = props;
 
   const isFetchingPage = useSelector(getIsFetchingPage);
-  const canvasWidth = useSelector(getCanvasWidth);
   const widgetsStructure = useSelector(getCanvasWidgetsStructure, equal);
   const pages = useSelector(getViewModePageList);
   const theme = useSelector(getCurrentThemeDetails);
@@ -143,11 +140,9 @@ function MainContainerWrapper(props: MainCanvasWrapperProps) {
   const isWDSV2Enabled = useFeatureFlag("ab_wds_enabled");
 
   const checkLayoutSystemFeatures = useLayoutSystemFeatures();
-  const [enableMainContainerResizer, enableOverlayCanvas] =
-    checkLayoutSystemFeatures([
-      LayoutSystemFeatures.ENABLE_MAIN_CONTAINER_RESIZER,
-      LayoutSystemFeatures.ENABLE_CANVAS_OVERLAY_FOR_EDITOR_UI,
-    ]);
+  const [enableMainContainerResizer] = checkLayoutSystemFeatures([
+    LayoutSystemFeatures.ENABLE_MAIN_CONTAINER_RESIZER,
+  ]);
 
   useEffect(() => {
     return () => {
@@ -174,7 +169,7 @@ function MainContainerWrapper(props: MainCanvasWrapperProps) {
   if (!isPageInitializing && widgetsStructure) {
     node = (
       <Canvas
-        canvasWidth={canvasWidth}
+        canvasWidth={props.canvasWidth}
         enableMainCanvasResizer={enableMainContainerResizer}
         pageId={params.pageId}
         widgetsStructure={widgetsStructure}
@@ -254,12 +249,6 @@ function MainContainerWrapper(props: MainCanvasWrapperProps) {
           </div>
         )}
         {node}
-        {enableOverlayCanvas && (
-          <OverlayCanvasContainer
-            canvasWidth={canvasWidth}
-            containerRef={props.parentRef}
-          />
-        )}
       </Wrapper>
       <MainContainerResizer
         currentPageId={currentPageId}
